Draw principal light rays
Draw the principal light rays (parallel ray, focal ray, and vertex ray)
starting from the object and extending off the canvas. If image is
virtual, extend rays backwards from the lens to meet image. Create a
function to draw a line given its endpoints, color, and whether is
should be dashed. Add a check to prevent placing the object on the right
side of the lens.