Add nonfunctional HTML input controls
Add a nonfunctional control panel area to the top-left of the screen to
display information and change parameters. Add numeric input controls to
change the object distance, object height, and the focal length of the
lens. Add a title for the simulation, a link to the GitHub repository,
and an area for usage instructions.