Update simulation based on input controls
Redraw the simulation with a new value for object distance, object
height, or focal length when the respective HTML input control fires a
change event. Calculate and display the image distance and height. If
the object moves as a result of a click event on the canvas, update the
values of the input controls.